refactor(eval): move SWEBenchInfo from logmanager to eval/swebench/info.py
Per Erik's review: SWEBenchInfo doesn't belong in logmanager.py, move it
into the eval/swebench module where it belongs.

- New: gptme/eval/swebench/info.py with SWEBenchInfo dataclass
- Updated gptme/eval/swebench/__init__.py to export SWEBenchInfo
- Removed SWEBenchInfo from gptme/logmanager.py
- Updated all imports in eval/agents/swebench.py, swe_extra/{run_swe_extra,
  swe_bench_test_spec,swe_bench_extra_data}.py
